+++ eno-core/src/main/java/fr/insee/eno/core/processing/out/steps/lunatic/table/DynamicTableQuestionProcessing.java
@@ -23,17 +23,7 @@ public class DynamicTableQuestionProcessing {
         // Header
         lunaticRoster.getHeader().addAll(HeaderCellsProcessing.from(enoTable, 0));
 
-        LinesRoster lines = new LinesRoster();
-        LabelType minLabel = new LabelType();
-        minLabel.setType(LabelTypeEnum.VTL);
-        minLabel.setValue(Integer.toString(enoTable.getMinLines().intValue()));
-        lines.setMin(minLabel);
-
-        LabelType maxLabel = new LabelType();
-        maxLabel.setType(LabelTypeEnum.VTL);
-        maxLabel.setValue(Integer.toString(enoTable.getMaxLines().intValue()));
-        lines.setMax(maxLabel);
-        lunaticRoster.setLines(lines);
+        setRosterSize(lunaticRoster, enoTable);
 
         List<TableCell> enoTableCells = new ArrayList<>();
         enoTableCells.addAll(enoTable.getResponseCells());
@@ -46,4 +36,41 @@ public class DynamicTableQuestionProcessing {
             lunaticRoster.getComponents().add(lunaticCell);
         }
     }
+
+    private static void setRosterSize(RosterForLoop lunaticRoster, DynamicTableQuestion enoTable) {
+        LinesRoster lines = new LinesRoster();
+        if (enoTable.getMinLines() != null && enoTable.getMaxLines() != null)
+            setMinMaxProperties(enoTable, lines);
+        else if (enoTable.getSizeExpression() != null)
+            setSizeExpression(enoTable, lines);
+        else
+            throw new IllegalStateException(
+                    "Table question '" + enoTable.getId() + "' has neither a min/max nor an expression size.");
+        lunaticRoster.setLines(lines);
+    }
+
+    private static void setMinMaxProperties(DynamicTableQuestion enoTable, LinesRoster lines) {
+        LabelType minLabel = new LabelType();
+        minLabel.setType(LabelTypeEnum.VTL);
+        minLabel.setValue(enoTable.getMinLines().toString());
+        lines.setMin(minLabel);
+
+        LabelType maxLabel = new LabelType();
+        maxLabel.setType(LabelTypeEnum.VTL);
+        maxLabel.setValue(enoTable.getMaxLines().toString());
+        lines.setMax(maxLabel);
+    }
+
+    /** Business rule: if the size of the roster for loop (dynamic table) is defined by an expression, the expression
+     * cannot be different between min and max.
+     * Note: in this case, we could have chosen to create an 'iterations' property in the Lunatic model to have the
+     * same property as in loop objects, but the min=max solution does the job. */
+    private static void setSizeExpression(DynamicTableQuestion enoTable, LinesRoster lines) {
+        LabelType sizeLabel = new LabelType();
+        sizeLabel.setType(LabelTypeEnum.VTL);
+        sizeLabel.setValue(enoTable.getSizeExpression().getValue());
+        lines.setMin(sizeLabel);
+        lines.setMax(sizeLabel);
+    }
+
 }